+++ src/app/home/page.tsx
@@ -1,5 +1,5 @@
-import Logout from "@/lib/components/Logout";
-import { validateSession } from "@/lib/server/auth";
+import Logout from "@/components/Logout";
+import { validateSession } from "@/server/auth";
 import { redirect } from "next/navigation";
 
 export default async function page() {
