+++ src/app/signup/page.tsx
@@ -1,5 +1,5 @@
-import SignUp from '@/lib/components/SignUp'
-import { validateSession } from '@/lib/server/auth';
+import SignUp from '@/components/SignUp';
+import { validateSession } from '@/server/auth';
 import { redirect } from 'next/navigation';
 
 export default async function page() {
